requirements.txt
# local package

# external requirements

pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml
mlflow
joblib
argparse
fastapi
uvicorn
pydantic
# for testing
pytest
pytest-cov  